# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTrap Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateRoomBody {
  hostName: string;
  /**
   * @minimum 2
   * @maximum 10
   */
  maxPlayers?: number;
  /**
   * @minimum 5
   * @maximum 30
   */
  questionCount?: number;
}

export interface PlayerAbilities {
  /** Remaining uses of confuse ability */
  confuse: number;
  /** Remaining uses of freeze ability */
  freeze: number;
  /** Remaining uses of reverse ability */
  reverse: number;
}

export interface Player {
  name: string;
  score: number;
  isHost: boolean;
  abilities: PlayerAbilities;
}

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  waiting: "waiting",
  playing: "playing",
  finished: "finished",
} as const;

export interface Room {
  id: string;
  code: string;
  hostName: string;
  players: Player[];
  status: RoomStatus;
  maxPlayers: number;
  questionCount: number;
  currentQuestion: number;
}

export type QuestionCategory =
  (typeof QuestionCategory)[keyof typeof QuestionCategory];

export const QuestionCategory = {
  tricky: "tricky",
  psychological: "psychological",
  visual: "visual",
  reverse: "reverse",
} as const;

export interface Question {
  id: string;
  text: string;
  options: string[];
  correctIndex: number;
  category: QuestionCategory;
  /** Seconds to answer (5-8) */
  timeLimit: number;
  explanation: string;
}

export interface QuestionsResponse {
  questions: Question[];
  total: number;
}

export interface LeaderboardEntry {
  rank: number;
  name: string;
  score: number;
  correctAnswers: number;
  totalAnswers: number;
  avgResponseTime: number;
}

export interface LeaderboardResponse {
  roomCode: string;
  entries: LeaderboardEntry[];
}

export type PlayerStatsProfileType =
  (typeof PlayerStatsProfileType)[keyof typeof PlayerStatsProfileType];

export const PlayerStatsProfileType = {
  fast_reckless: "fast_reckless",
  slow_precise: "slow_precise",
  hard_to_trick: "hard_to_trick",
  easy_to_trick: "easy_to_trick",
} as const;

export interface PlayerStats {
  playerName: string;
  profileType: PlayerStatsProfileType;
  profileLabel: string;
  profileEmoji: string;
  score: number;
  rank: number;
  totalPlayers: number;
  correctAnswers: number;
  wrongAnswers: number;
  avgResponseTime: number;
  bonusPoints: number;
  abilitiesUsed: number;
}

export type GetQuestionsParams = {
  count?: number;
  category?: GetQuestionsCategory;
};

export type GetQuestionsCategory =
  (typeof GetQuestionsCategory)[keyof typeof GetQuestionsCategory];

export const GetQuestionsCategory = {
  tricky: "tricky",
  psychological: "psychological",
  visual: "visual",
  reverse: "reverse",
  all: "all",
} as const;