# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTrap Game API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new game room
 */
export const createRoomBodyMaxPlayersDefault = 6;
export const createRoomBodyMaxPlayersMin = 2;
export const createRoomBodyMaxPlayersMax = 10;

export const createRoomBodyQuestionCountDefault = 10;
export const createRoomBodyQuestionCountMin = 5;
export const createRoomBodyQuestionCountMax = 30;

export const CreateRoomBody = zod.object({
  hostName: zod.string(),
  maxPlayers: zod
    .number()
    .min(createRoomBodyMaxPlayersMin)
    .max(createRoomBodyMaxPlayersMax)
    .default(createRoomBodyMaxPlayersDefault),
  questionCount: zod
    .number()
    .min(createRoomBodyQuestionCountMin)
    .max(createRoomBodyQuestionCountMax)
    .default(createRoomBodyQuestionCountDefault),
});

/**
 * @summary Get room by code
 */
export const GetRoomParams = zod.object({
  code: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  id: zod.string(),
  code: zod.string(),
  hostName: zod.string(),
  players: zod.array(
    zod.object({
      name: zod.string(),
      score: zod.number(),
      isHost: zod.boolean(),
      abilities: zod.object({
        confuse: zod.number().describe("Remaining uses of confuse ability"),
        freeze: zod.number().describe("Remaining uses of freeze ability"),
        reverse: zod.number().describe("Remaining uses of reverse ability"),
      }),
    }),
  ),
  status: zod.enum(["waiting", "playing", "finished"]),
  maxPlayers: zod.number(),
  questionCount: zod.number(),
  currentQuestion: zod.number(),
});

/**
 * @summary Get questions for the game
 */
export const getQuestionsQueryCountDefault = 10;
export const getQuestionsQueryCategoryDefault = `all`;

export const GetQuestionsQueryParams = zod.object({
  count: zod.coerce.number().default(getQuestionsQueryCountDefault),
  category: zod
    .enum(["tricky", "psychological", "visual", "reverse", "all"])
    .default(getQuestionsQueryCategoryDefault),
});

export const GetQuestionsResponse = zod.object({
  questions: zod.array(
    zod.object({
      id: zod.string(),
      text: zod.string(),
      options: zod.array(zod.string()),
      correctIndex: zod.number(),
      category: zod.enum(["tricky", "psychological", "visual", "reverse"]),
      timeLimit: zod.number().describe("Seconds to answer (5-8)"),
      explanation: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get final leaderboard for a room
 */
export const GetLeaderboardParams = zod.object({
  roomCode: zod.coerce.string(),
});

export const GetLeaderboardResponse = zod.object({
  roomCode: zod.string(),
  entries: zod.array(
    zod.object({
      rank: zod.number(),
      name: zod.string(),
      score: zod.number(),
      correctAnswers: zod.number(),
      totalAnswers: zod.number(),
      avgResponseTime: zod.number(),
    }),
  ),
});

/**
 * @summary Get player analytics after game
 */
export const GetPlayerStatsParams = zod.object({
  roomCode: zod.coerce.string(),
  playerName: zod.coerce.string(),
});

export const GetPlayerStatsResponse = zod.object({
  playerName: zod.string(),
  profileType: zod.enum([
    "fast_reckless",
    "slow_precise",
    "hard_to_trick",
    "easy_to_trick",
  ]),
  profileLabel: zod.string(),
  profileEmoji: zod.string(),
  score: zod.number(),
  rank: zod.number(),
  totalPlayers: zod.number(),
  correctAnswers: zod.number(),
  wrongAnswers: zod.number(),
  avgResponseTime: zod.number(),
  bonusPoints: zod.number(),
  abilitiesUsed: zod.number(),
});